+++ resources/views/admin/sites/partials/theme-tab.blade.php
@@ -1,0 +1,1 @@
+@include('webblocks-cms::admin.sites.partials.theme-tab', get_defined_vars())